+++ src/main/kotlin/racingcar/domain/Rank.kt
@@ -1,11 +1,11 @@
-package racingcar
+package racingcar.domain
 
 object Rank {
     fun getWinnerNames(cars: List<Car>): String =
         this.getWinner(cars)
             .joinToString(",") { it.name }
 
-    private fun getWinner(cars: List<Car>): List<Car> {
+    fun getWinner(cars: List<Car>): List<Car> {
         val carsByPosition = cars.groupBy { it.position }
         val maxPosition: Int = carsByPosition.maxOf { it.key }
